+++ src/Migration/Migration1710243839IncludeHiddenCategories.php
@@ -1,0 +1,27 @@
+<?php declare(strict_types=1);
+
+namespace RH\Tweakwise\Migration;
+
+use Doctrine\DBAL\Connection;
+use Shopware\Core\Framework\Migration\MigrationStep;
+
+class Migration1710243839IncludeHiddenCategories extends MigrationStep
+{
+    public function getCreationTimestamp(): int
+    {
+        return 1710243839;
+    }
+
+    public function update(Connection $connection): void
+    {
+        $sql = <<<SQL
+ALTER TABLE `s_plugin_rhae_tweakwise_feed`
+                ADD `includeHiddenCategories` TINYINT(1) NOT NULL DEFAULT '0';
+SQL;
+        $connection->executeUpdate($sql);
+    }
+
+    public function updateDestructive(Connection $connection): void
+    {
+    }
+}