+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Added
+- In the feed configuration you can now define if categories which are hidden in the navigation should be exported for Tweakwise or not.
+
 ## [2.5.0] - 12-03-2024
 
 ### Added